requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
Pillow
django-location-field